Add files via upload
This study's anticipated results are two-fold: First, we anticipate the creation of an entity-level sentiment analysis model that is reliable and accurate that has been specially designed for Twitter data. This model will make it possible to identify sentiments in connection with entities in a tweet, giving us a more detailed knowledge of the feelings conveyed. The development of a user-friendly tool or application that can be used by individuals, corporations, and organizations to extract useful information from Twitter data is our second goal. With the help of this tool, users will be better able to monitor and respond to public opinion surrounding the entities that interest them, which will lead to increased participation and better decision-making.